- fixed incorrect rounding in several places in the code - diagonal movement now uses the proper amount of AP and BP
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@8634 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Laptop/personnel.cpp
@@ -8986,7 +8986,7 @@ INT8 CalculateMercsAchievementPercentage( INT32 ubProfile )
 	// Calculate percentage
 	if( ulTotalMercPoints != 0 )
 	{
-		uiMercPercentage = (UINT32)(((100 * uiMercPoints)/ulTotalMercPoints)+ 0.5);
+		uiMercPercentage = (UINT32)(100.0f * (FLOAT)uiMercPoints / (FLOAT)ulTotalMercPoints + 0.5f);
 
 		uiMercPercentage = min( 100, uiMercPercentage );
